// data/foodb-models/db.d.ts
export interface foodsAttribute {
  id:number;
  name:string;
  nameScientific?:string;
  description?:string;
  itisId?:string;
  wikipediaId?:string;
  pictureFileName?:string;
  pictureContentType?:string;
  pictureFileSize?:number;
  pictureUpdatedAt?:Date;
  legacyId?:number;
  foodGroup?:string;
  foodSubgroup?:string;
  foodType:string;
  createdAt?:Date;
  updatedAt?:Date;
  creatorId?:number;
  updaterId?:number;
  exportToAfcdb:number;
  category?:string;
  ncbiTaxonomyId?:number;
  exportToFoodb?:number;
}
export interface foodsInstance extends Sequelize.Instance<foodsAttribute>, foodsAttribute { }